+++ src/app/jobs/(jobs-category)/public-jobs/page.tsx
@@ -1,9 +1,9 @@
 "use client";
 import SelectedChips from "@/features/jobs/components/SelectedChips";
+import { useFilterTabStore } from "@/features/jobs/stores/job-filters/useJobFilterTabsStore";
 import { useEffect } from "react";
 import JobCard from "../../../../features/home/components/JobCard";
 import JobFilter from "../../../../features/jobs/components/JobFilter";
-import { useFilterTabStore } from "../../../../stores/useJobFilterStore";
 
 export default function PublicJobsPage() {
   const setShowLocation = useFilterTabStore((state) => state.setShowLocation);
