+++ src/features/jobs/components/JobFilter.tsx
@@ -2,7 +2,8 @@
 
 import { Heading } from "@/components/ui/Heading";
 import { FaCaretDown } from "react-icons/fa";
-import { useFilterTabStore } from "../../../stores/useJobFilterStore";
+
+import { useFilterTabStore } from "@/features/jobs/stores/job-filters/useJobFilterTabsStore";
 import FilterJobs from "../filter/JobCategoryFilter";
 import FilterOtherConditions from "../filter/JobConditionsFilter";
 import FilterLocation from "../filter/JobLocationFilter";
