+++ src/features/jobs/stores/job-filters/useSelectedFiltersStore.ts
@@ -1,39 +1,9 @@
+import { REGIONS } from "@/constants/regions";
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
-import { REGIONS } from "../constants/regions";
 
-type FilterTabStore = {
-  showLocation: boolean;
-  showJobs: boolean;
-  showOtherConditions: boolean;
-  setShowLocation: (value: boolean) => void;
-  setShowJobs: (value: boolean) => void;
-  setShowOtherConditions: (value: boolean) => void;
-};
 
-export const useFilterTabStore = create<FilterTabStore>((set) => ({
-  showLocation: false,
-  showJobs: false,
-  showOtherConditions: false,
-  setShowLocation: (value) =>
-    set(() => ({
-      showLocation: value,
-      showJobs: false,
-      showOtherConditions: false,
-    })),
-  setShowJobs: (value) =>
-    set(() => ({
-      showJobs: value,
-      showLocation: false,
-      showOtherConditions: false,
-    })),
-  setShowOtherConditions: (value) =>
-    set(() => ({
-      showOtherConditions: value,
-      showLocation: false,
-      showJobs: false,
-    })),
-}));
+
 
 type FilterSelectedStore = {
   selectedFilters: string[];
